requirements.txt
# API Framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
geoalchemy2>=0.13.0
alembic>=1.12.0

# Geospatial
geopandas>=0.13.0
shapely>=2.0.0
pyproj>=3.6.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
pulp>=2.7.0
ortools>=9.6.0

# ML Tracking
mlflow>=2.8.0
wandb>=0.15.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=1.1.3
fastapi-mail>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Monitoring
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
prometheus-client>=0.17.0

# Development
black>=23.7.0
isort>=5.12.0
mypy>=1.5.0
flake8>=6.1.0

# Visualization and Export
xlsxwriter>=3.0.0
python-dotenv>=0.19.0 